# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Deck OS - JARVIS Cyberdeck Command Center API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /plugins/{pluginId}:
    get:
      operationId: getPlugin
      tags: [plugins]
      summary: Get plugin details
      parameters:
        - name: pluginId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Plugin details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Plugin"
        "404":
          description: Plugin not found
  /plugins/{pluginId}/toggle:
    post:
      operationId: togglePlugin
      tags: [plugins]
      summary: Enable or disable a plugin
      parameters:
        - name: pluginId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TogglePluginRequest"
      responses:
        "200":
          description: Updated plugin state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Plugin"
  /plugins/{pluginId}/execute:
    post:
      operationId: executePluginCommand
      tags: [plugins]
      summary: Execute a command on a plugin
      parameters:
        - name: pluginId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PluginCommandRequest"
      responses:
        "200":
          description: Command result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PluginCommandResult"
...
components:
  schemas:
...
    Plugin:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        version:
          type: string
        enabled:
          type: boolean
        status:
          type: string
          enum: [active, inactive, error, loading]
        commands:
          type: array
          items:
            type: string
        category:
          type: string
        lastActivity:
          type: string
          format: date-time
          nullable: true
        errorMessage:
          type: string
          nullable: true
      required:
        - id
        - name
        - description
        - version
        - enabled
        - status
        - commands
        - category
    TogglePluginRequest:
      type: object
      properties:
        enabled:
          type: boolean
      required:
        - enabled
    PluginCommandRequest:
      type: object
      properties:
        command:
          type: string
        args:
          type: object
          additionalProperties: true
        mode:
          type: string
          enum: [fast, deep, none]
          default: none
      required:
        - command
    PluginCommandResult:
      type: object
      properties:
        success:
          type: boolean
        output:
          type: string
        data:
          type: object
          additionalProperties: true
          nullable: true
        executionTimeMs:
          type: integer
        modelUsed:
          type: string
          nullable: true
      required:
        - success
        - output
        - executionTimeMs
...